% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcoderank.R
\name{get_lsoa}
\alias{get_lsoa}
\title{Get lsoa numbers}
\usage{
get_lsoa(postcode_info)
}
\arguments{
\item{postcode_info}{Postcode information datasets}
}
\description{
Utility function to pull LSOA numbers from the postcode_info dataset
}
