% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcoderank.R
\name{get_lsoa_scores}
\alias{get_lsoa_scores}
\title{Get lsoa scores}
\usage{
get_lsoa_scores(LSOA, mdi)
}
\arguments{
\item{LSOA}{LSOA number}

\item{data}{MDI source dataset}
}
\description{
Utility funciton to extract area scores from the MDI dataset based upon a LSOA number
}
