% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_mdi_data}
\alias{get_mdi_data}
\title{Get multiple deprivation indicies data}
\usage{
get_mdi_data()
}
\arguments{
\item{infile}{name of local csv file to save the data to}
}
\description{
Makes a local copy of the government provided indices of deprivation
}
