% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postcode_info.R
\name{get_postcode_info}
\alias{get_postcode_info}
\title{Get postcode lookup info}
\usage{
get_postcode_info(postcode)
}
\arguments{
\item{postcode}{Single UK postcode i.e. "EC3N 4AB"}
}
\description{
Function takes a postcode and returns standard lookup information. Note that spaces are automatically removed.
Information is sourced from https://www.doogal.co.uk/ thus an internet connection is required to use this function.
}
