% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcoderank.R
\docType{package}
\name{postcodeRank-package}
\alias{postcodeRank}
\alias{postcodeRank-package}
\title{postcodeRank: Support functions to lookup multiple deprivation indicies information of various UK postcodes}
\description{
Support functions to lookup multiple deprivation indicies information of various UK postcodes
}
\author{
\strong{Maintainer}: Craig Gower \email{first.last@example.com}

}
\keyword{internal}
