% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcoderank.R
\name{postcodeRank}
\alias{postcodeRank}
\title{Get the MDI scores for various postcodes}
\usage{
postcodeRank(postcodes, mdi)
}
\arguments{
\item{postcodes}{character vector of postcodes}

\item{mdi}{MDI dataset (use get_mdi_data() if you don't have one already downloaded)}
}
\description{
Get the MDI scores for various postcodes
}
