% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcoderank.R
\name{postcodeRank_full}
\alias{postcodeRank_full}
\title{Get postcore score information}
\usage{
postcodeRank_full(postcode, mdi)
}
\arguments{
\item{postcode}{UK Postcode}
}
\description{
Returns the MDI scores for a given postcode based upon its LSOA region
}
